separo app de su construcción
